backend/requirements.txt
Django==6.0.3
django-environ==0.13.0
djangorestframework==3.17.0
djangorestframework-simplejwt==5.5.1

celery==5.6.2
redis==7.3.0
kombu==5.6.2
billiard==4.2.4
vine==5.1.0
amqp==5.3.1

psycopg2-binary==2.9.11
pillow==12.1.1
requests==2.32.5
python-dotenv==1.2.2

asgiref==3.11.1
sqlparse==0.5.5
tzdata==2025.3
tzlocal==5.3.1

certifi==2026.2.25
charset-normalizer==3.4.6
idna==3.11
urllib3==2.6.3

click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
prompt-toolkit==3.0.52
wcwidth==0.6.0
packaging==26.0
python-dateutil==2.9.0.post0
six==1.17.0
PyJWT==2.12.1
django-environ==0.13.0
pgvector==0.3.6
beautifulsoup4
lxml
